fix: migrations

+++ database/migrations/2024_12_11_000009_create_command_results_table.php
@@ -8,18 +8,20 @@ return new class() extends Migration
 {
     public function up(): void
     {
-        Schema::create('command_results', function (Blueprint $table) {
-            $table->id();
-            $table->boolean('success');
-            $table->string('command');
-            $table->longText('response')->nullable();
-            $table->text('error')->nullable();
-            $table->json('result')->nullable();
-            $table->timestamp('created_at')->useCurrent();
+        if (! Schema::hasTable('command_results')) {
+            Schema::create('command_results', function (Blueprint $table) {
+                $table->id();
+                $table->boolean('success');
+                $table->string('command');
+                $table->longText('response')->nullable();
+                $table->text('error')->nullable();
+                $table->json('result')->nullable();
+                $table->timestamp('created_at')->useCurrent();
 
-            $table->unsignedBigInteger('batch_id')->nullable()->index();
-            $table->foreign('batch_id')->references('id')->on('command_result_batches')->onDelete('cascade');
-        });
+                $table->unsignedBigInteger('batch_id')->nullable()->index();
+                $table->foreign('batch_id')->references('id')->on('command_result_batches')->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
